Update README.md for instructions for Gatsby
I found this library on the official Gatsby page for implementing the comments section. This project is awesome.

I faced an error like this when I used this library in my Gatsby powered website:

`Error: Validation Failed. The listed users and repositories cannot be searched either because the resources do not exist or you do not have permission to view them.`

I used the URL form of repo in options. It is not explicitly stated in README to use the non-URL form of the repository.

Also, the `gatsby-plugin-gitalk` helped me to completely implement this project in Gatsby. A huge shoutout to @suziwen for creating this plugin.

I don't want other developers to find this dilemma like mine. So I updated the `README.md` file to include the necessary instructions for Gatsby developers.